rename Telemetry to ProductTelemetry
