+++ .gitignore
@@ -14,7 +14,9 @@ index_data
 # Default configuration for persist_directory in chromadb/config.py
 # Currently it's located in "./chroma/"
 chroma/
+chroma_test_data/
 
+.venv
 venv
 .env
 .chroma
